requirements.txt
# Core Dependencies
pandas>=1.3.0
numpy>=1.20.0
requests>=2.28.2
matplotlib>=3.5.0 # For plotting results
seaborn>=0.11.0 # Optional: For enhanced plot styling

# API Integration & Data Fetching
python-graphql-client>=0.4.0 # For interacting with TheGraph subgraphs
python-dotenv>=0.19.0 # For loading environment variables (.env file)

# CLI & Formatting
# argparse (built-in)

# Configuration
PyYAML>=5.4 # If using YAML for config (config.py uses it optionally)

# Testing (Optional: For development)
# pytest>=7.3.0